Use C strings wherever reasonable
Given that the inputs to the generator are the command line and TinyXML which
give C strings using std::string everywhere is going against the grain as well
as slowing down compile and run times with no clear benefit. Keep using std::string
for the indent string and type declaration text as these applications take advantage
of the fact that std::string's store the string length.