Remove GLX_EXTENSION_NAME from glx.xml
Reluctantly hard coding this part of the specs into the generator. It
tries to assign an enum with a string in order to trick the generator to
outputing the string as a #define. This is inconvienent for glbindify
since it stores an integer value for each enum in the first pass.