Changes and fixes to support EGL
- Support string enum values. EGL defines some enums with typecasts in
them so the numeric map cannot be used. This provides a cleaner fix for
the problem with GLX_EXTENSION_NAME as well.

- Store types unassociated with any feature version or extension in a
separate structure. These types are printed unconditionally before any
other types. This ensures that dependencies are met properly
as the dependency checking logic was unable to recognize dependencies
occuring in an EGL function pointer type.

- Use eglGetProcAddress for the EGL API and support both glxGetProcAddress
and eglGetProcAddress for GL through a GLB_USE_EGL macro